+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/dumpview/DumpInfo.java
@@ -124,7 +124,7 @@ public class DumpInfo implements TreeItem {
     @Override
     public String toString() {
         String value = previewValue == null ? "" : previewValue.toString();
-        return name + " (" + type + ")" + (value.isEmpty() ? "" : " = " + value);
+        return name + " : " + type + (value.isEmpty() ? "" : " = " + value);
     }
 
     public void resolveTag() {